+++ NumeRe.cbp
@@ -269,8 +269,6 @@
 		<Unit filename="common/filerevisions.hpp" />
 		<Unit filename="common/filewatcher.cpp" />
 		<Unit filename="common/filewatcher.hpp" />
-		<Unit filename="common/http.cpp" />
-		<Unit filename="common/http.h" />
 		<Unit filename="common/ipc.cpp" />
 		<Unit filename="common/ipc.hpp" />
 		<Unit filename="common/options.cpp" />
@@ -415,11 +413,6 @@
 		<Unit filename="kernel/core/AudioLib/audiofile.hpp" />
 		<Unit filename="kernel/core/AudioLib/wavfile.cpp" />
 		<Unit filename="kernel/core/AudioLib/wavfile.hpp" />
-		<Unit filename="kernel/core/IgorLib/CrossPlatformFileIO.cpp" />
-		<Unit filename="kernel/core/IgorLib/CrossPlatformFileIO.h" />
-		<Unit filename="kernel/core/IgorLib/IgorBin.h" />
-		<Unit filename="kernel/core/IgorLib/ReadWave.cpp" />
-		<Unit filename="kernel/core/IgorLib/ReadWave.h" />
 		<Unit filename="kernel/core/ParserLib/muParser.cpp" />
 		<Unit filename="kernel/core/ParserLib/muParser.h" />
 		<Unit filename="kernel/core/ParserLib/muParserBase.cpp" />
@@ -484,6 +477,15 @@
 		<Unit filename="kernel/core/documentation/documentation.hpp" />
 		<Unit filename="kernel/core/interval.cpp" />
 		<Unit filename="kernel/core/interval.hpp" />
+		<Unit filename="kernel/core/io/BasicExcel.cpp" />
+		<Unit filename="kernel/core/io/BasicExcel.hpp" />
+		<Unit filename="kernel/core/io/IgorLib/CrossPlatformFileIO.cpp" />
+		<Unit filename="kernel/core/io/IgorLib/CrossPlatformFileIO.h" />
+		<Unit filename="kernel/core/io/IgorLib/IgorBin.h" />
+		<Unit filename="kernel/core/io/IgorLib/ReadWave.cpp" />
+		<Unit filename="kernel/core/io/IgorLib/ReadWave.h" />
+		<Unit filename="kernel/core/io/archive.cpp" />
+		<Unit filename="kernel/core/io/archive.hpp" />
 		<Unit filename="kernel/core/io/file.cpp" />
 		<Unit filename="kernel/core/io/file.hpp" />
 		<Unit filename="kernel/core/io/fileops.cpp" />
@@ -496,6 +498,12 @@
 		<Unit filename="kernel/core/io/output.hpp" />
 		<Unit filename="kernel/core/io/styledtextfile.cpp" />
 		<Unit filename="kernel/core/io/styledtextfile.hpp" />
+		<Unit filename="kernel/core/io/unzip.cpp" />
+		<Unit filename="kernel/core/io/unzip.h" />
+		<Unit filename="kernel/core/io/zip++.cpp" />
+		<Unit filename="kernel/core/io/zip++.hpp" />
+		<Unit filename="kernel/core/io/zip.cpp" />
+		<Unit filename="kernel/core/io/zip.h" />
 		<Unit filename="kernel/core/maths/command_implementations.cpp" />
 		<Unit filename="kernel/core/maths/command_implementations.hpp" />
 		<Unit filename="kernel/core/maths/define.cpp" />
@@ -580,14 +588,12 @@
 		<Unit filename="kernel/core/ui/language.hpp" />
 		<Unit filename="kernel/core/ui/winlayout.cpp" />
 		<Unit filename="kernel/core/ui/winlayout.hpp" />
-		<Unit filename="kernel/core/utils/BasicExcel.cpp" />
-		<Unit filename="kernel/core/utils/BasicExcel.hpp" />
-		<Unit filename="kernel/core/utils/archive.cpp" />
-		<Unit filename="kernel/core/utils/archive.hpp" />
 		<Unit filename="kernel/core/utils/datetimetools.cpp" />
 		<Unit filename="kernel/core/utils/datetimetools.hpp" />
 		<Unit filename="kernel/core/utils/filecheck.cpp" />
 		<Unit filename="kernel/core/utils/filecheck.hpp" />
+		<Unit filename="kernel/core/utils/http.cpp" />
+		<Unit filename="kernel/core/utils/http.h" />
 		<Unit filename="kernel/core/utils/stringtools.cpp" />
 		<Unit filename="kernel/core/utils/stringtools.hpp" />
 		<Unit filename="kernel/core/utils/timer.hpp">
@@ -597,12 +603,6 @@
 		</Unit>
 		<Unit filename="kernel/core/utils/tools.cpp" />
 		<Unit filename="kernel/core/utils/tools.hpp" />
-		<Unit filename="kernel/core/utils/unzip.cpp" />
-		<Unit filename="kernel/core/utils/unzip.h" />
-		<Unit filename="kernel/core/utils/zip++.cpp" />
-		<Unit filename="kernel/core/utils/zip++.hpp" />
-		<Unit filename="kernel/core/utils/zip.cpp" />
-		<Unit filename="kernel/core/utils/zip.h" />
 		<Unit filename="kernel/core/version.h" />
 		<Unit filename="kernel/kernel.cpp" />
 		<Unit filename="kernel/kernel.hpp" />
