+++ kernel/core/datamanagement/dataops.cpp
@@ -22,7 +22,6 @@
 #include "../../kernel.hpp"
 #include "../ui/language.hpp"
 #include "../utils/tools.hpp"
-#include "../utils/BasicExcel.hpp"
 #include "../ui/error.hpp"
 #include "../structures.hpp"
 #include "../built-in.hpp"
