+++ kernel/core/io/IgorLib/ReadWave.cpp
